+++ .eslintrc.cjs
@@ -1,9 +1,35 @@
 module.exports = {
   parser: "@typescript-eslint/parser", // Specifies the ESLint parser
-  plugins: ["unicorn", "turbo"],
+  plugins: ["turbo"],
   extends: [
-    "plugin:@typescript-eslint/recommended", // Uses the recommended rules from the @typescript-eslint/eslint-plugin
+    "plugin:@typescript-eslint/recommended",
     "plugin:prettier/recommended",
+  ],
+  overrides: [
+    {
+      extends: [
+        "plugin:@typescript-eslint/recommended-requiring-type-checking",
+      ],
+      files: ["*.ts", "*.tsx"],
+      parserOptions: {
+        project: "tsconfig.json",
+      },
+      rules: {
+        "@typescript-eslint/require-await": "off",
+        "@typescript-eslint/restrict-plus-operands": "off",
+        "@typescript-eslint/restrict-template-expressions": "off",
+
+        // These rules are only disabled because we hit a bug in linting.
+        // See https://github.com/t3-oss/create-t3-app/pull/1036#discussion_r1060505136
+        // If you still see the bug once TypeScript@5 is used, please let typescript-eslint know!
+        "@typescript-eslint/no-unsafe-argument": "off",
+        "@typescript-eslint/no-unsafe-assignment": "off",
+        "@typescript-eslint/no-unsafe-call": "off",
+        "@typescript-eslint/no-unsafe-call": "off",
+        "@typescript-eslint/no-unsafe-member-access": "off",
+        "@typescript-eslint/no-unsafe-return": "off",
+      },
+    },
   ],
   parserOptions: {
     ecmaVersion: "latest", // Allows for the parsing of modern ECMAScript features
@@ -12,16 +38,11 @@ module.exports = {
   },
   rules: {
     // Place to specify ESLint rules. Can be used to overwrite rules specified from the extended configs
-    "@typescript-eslint/explicit-function-return-type": "off",
-    "@typescript-eslint/explicit-module-boundary-types": "off",
     "@typescript-eslint/no-explicit-any": "off",
-    "@typescript-eslint/no-floating-promises": "error",
-    "unicorn/filename-case": [
+    "@typescript-eslint/no-non-null-assertion": "off",
+    "@typescript-eslint/no-unused-vars": [
       "error",
-      {
-        case: "camelCase",
-      },
+      { argsIgnorePattern: "^_", destructuredArrayIgnorePattern: "^_" },
     ],
-    "prettier/prettier": ["error", { endOfLine: "auto" }],
   },
 };