+++ www/.eslintrc.cjs
@@ -35,22 +35,10 @@ module.exports = {
     {
       files: ["*.ts", "*.tsx"],
       parser: "@typescript-eslint/parser",
-      extends: ["plugin:@typescript-eslint/recommended"],
       rules: {
-        "@typescript-eslint/no-unused-vars": [
+        "no-unused-vars": [
           "error",
           { argsIgnorePattern: "^_", destructuredArrayIgnorePattern: "^_" },
-        ],
-        "@typescript-eslint/no-non-null-assertion": "off",
-        "unicorn/filename-case": [
-          "error",
-          {
-            cases: {
-              camelCase: true,
-              // FIXME: Temporarily Accepting pascalCase for React files, see https://github.com/withastro/astro/issues/4220.
-              pascalCase: true,
-            },
-          },
         ],
       },
     },